backend/requirements.txt
flask
flask-cors
flask-jwt-extended
bcrypt
pymysql
python-dotenv
cryptography
gunicorn